backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
numpy==1.26.4
scipy==1.14.1
scikit-image==0.24.0
scikit-learn==1.5.2
opencv-python==4.10.0.84
pillow==11.0.0
torch==2.6.0
torchvision==0.21.0
timm==1.0.12
monai==1.5.1
matplotlib==3.9.3
albumentations==1.4.23
tqdm==4.67.1
python-dotenv==1.0.1
